refactored methods and functions
