+++ Week5/IMDB_project/app/src/main/java/com/example/imdb_project/data/repositories/MoviesRepositoryImpl.kt
@@ -2,25 +2,14 @@ package com.example.imdb_project.data.repositories
 
 import com.example.imdb_project.domain.models.MovieModel
 import com.example.imdb_project.data.datasources.MoviesDataSource
+import com.example.imdb_project.data.mappers.MovieListMapper
 
 // The repository is in charge of select the datasource (local or cloud)
 // The repository is in charge of executing transformations of data
 class MoviesRepositoryImpl(private val dataSource: MoviesDataSource): MoviesRepository {
     override fun getMovies(): List<MovieModel> {
         val dto = dataSource.getMovies()
-        val movies = mutableListOf<MovieModel>()
-        dto.forEach{ movieDto ->
-            movies.add( MovieModel(
-                title = movieDto.title,
-                description = movieDto.description,
-                thumbnail = movieDto.thumbnail,
-                preview = movieDto.preview,
-                actors = movieDto.actors,
-                rating = movieDto.rating,
-                numberOfEpisodes = movieDto.numberOfEpisodes,
-                year = movieDto.year
-            ))
-        }
-        return movies
+        val mapper = MovieListMapper()
+        return mapper.mapMovieDtoList2MovieModelList(dto)
     }
 }
