+++ Week5/IMDB_project/app/src/main/java/com/example/imdb_project/ui/presentation/fragments/LoginActivity.kt
@@ -1,9 +1,10 @@
-package com.example.imdb_project
+package com.example.imdb_project.ui.presentation.fragments
 
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
+import com.example.imdb_project.R
 import com.example.imdb_project.ui.presentation.HomeActivity
 
 class LoginActivity : AppCompatActivity(){
